Storing Components in Files and Using a good project structure
